pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = "Automação do ticket de alimentação"
authors = ["pferrariog <pdrferrari08@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
pandas = "^1.4.4"
openpyxl = "^3.0.10"
numpy = "^1.23.3"
requests = "^2.28.1"
selenium = "^4.4.3"
python-dotenv = "^0.21.0"
isort = "^5.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
